move `FeedResponse` and `FeedImageResponseMapper` into different files
